Add liked_posts field to Customer model
This commit adds a new field, `liked_posts`, to the `Customer` model in the `social` app. The field is a `ManyToManyField` that establishes a relationship between the `Customer` model and the `Post` model through the `Like` model. This allows customers to like multiple posts and retrieve the posts that they have liked.